% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outliers.R
\name{plot_outliers}
\alias{plot_outliers}
\title{Plot Pairwise Outliers}
\usage{
plot_outliers(data, method = c("mahalanobis", "mcd"), alpha = 0.975)
}
\arguments{
\item{data}{A numeric data frame or matrix.}

\item{method}{Outlier detection method: "mahalanobis" or "mcd".}

\item{alpha}{The quantile cutoff for identifying outliers (default 0.975).}
}
\description{
Generates 2D scatterplots for each pair of variables in the dataset,
with outliers identified using Mahalanobis or MCD distances
computed across all variables, without including each observation
in its own distance calculation.
}
\examples{
df_mtcars <- mtcars[, c("mpg", "hp", "wt" )]
head(df_mtcars)

## Pairwise Plots: Mahalanobis
plot_outliers(df_mtcars, method = "mahalanobis", alpha = 0.975)

## Pairwise Plots: MCD
plot_outliers(df_mtcars, method = "mcd", alpha = 0.975)
}
